requirements.txt
playwright
pandas
anthropic
python-dotenv
reportlab
schedule